Extract LogFormatProxyModel to another file
